Add onDelete function to DeleteTweetAlert component
